requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1

SQLAlchemy==2.0.45
alembic==1.17.2
asyncpg==0.31.0
psycopg-binary==3.3.2
psycopg==3.3.2